correcting the component_type type to boost::int32_t
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@6549 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ hpx/runtime/naming/address.hpp
@@ -29,7 +29,7 @@ namespace hpx { namespace naming
     ///
     struct HPX_EXPORT address
     {
-        typedef boost::int64_t component_type;
+        typedef boost::int32_t component_type;
         typedef boost::uint64_t address_type;
 
         ///////////////////////////////////////////////////////////////////////
